+++ src/main/java/emi/mtg/deckbuilder/view/dialogs/PrintingSelectorDialog.java
@@ -6,6 +6,7 @@ import emi.mtg.deckbuilder.model.Preferences;
 import emi.mtg.deckbuilder.view.components.CardPane;
 import emi.mtg.deckbuilder.view.layouts.FlowGrid;
 import javafx.application.Platform;
+import javafx.beans.binding.Bindings;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.scene.Scene;
@@ -28,12 +29,19 @@ public class PrintingSelectorDialog extends Dialog<Card.Printing> {
 
 		setTitle("Printing Selector");
 		setResult(null);
-		setResultConverter(bt -> getResult());
 
 		ObservableList<CardInstance> tmpModel = FXCollections.observableList(card.printings().stream()
 				.map(CardInstance::new)
 				.collect(Collectors.toList()));
 		pane = new CardPane("Variations", tmpModel, FlowGrid.Factory.INSTANCE);
+
+		setResultConverter(bt -> {
+			if (bt != ButtonType.CANCEL && pane.view().selectedCards.size() == 1) {
+				return pane.view().selectedCards.iterator().next().printing();
+			} else {
+				return null;
+			}
+		});
 
 		pane.view().doubleClick(ci -> {
 			setResult(ci.printing());
@@ -41,8 +49,10 @@ public class PrintingSelectorDialog extends Dialog<Card.Printing> {
 		});
 
 		pane.autoAction.set(ci -> {
-			setResult(ci.printing());
-			Platform.runLater(PrintingSelectorDialog.this::close);
+			Platform.runLater(() -> {
+				setResult(ci.printing());
+				close();
+			});
 		});
 
 		pane.autoEnabled.set(true);
@@ -54,7 +64,8 @@ public class PrintingSelectorDialog extends Dialog<Card.Printing> {
 
 		getDialogPane().setStyle(Preferences.get().theme.style());
 		getDialogPane().setContent(pane);
-		getDialogPane().getButtonTypes().add(ButtonType.CANCEL);
+		getDialogPane().getButtonTypes().setAll(ButtonType.OK, ButtonType.CANCEL);
+		getDialogPane().lookupButton(ButtonType.OK).disableProperty().bind(Bindings.size(pane.view().selectedCards).isNotEqualTo(1));
 		initModality(Modality.WINDOW_MODAL);
 		initOwner(scene.getWindow());